+++ jobs/etl_jobs/external/ppg/.gitignore
@@ -1,2 +1,1 @@
 reports/
-templates/
