+++ jobs/etl_jobs/external/ppg/core.py
@@ -2,626 +2,508 @@ from dataclasses import dataclass, field
 from pathlib import Path
 from typing import List, Dict, Optional, Any
 from enum import Enum
+from collections import defaultdict
 import duckdb
-import openpyxl
-import typer
-from google.cloud import bigquery
 import warnings
 
-from configs import GCP_PROJECT, BIGQUERY_PPG_DATASET, SCOPE_TABLES, QUERIES, REPORTS, TOP_TABLES_BY_SCOPE, STAKEHOLDER_REPORTS, ALL_TABLES,TAB_NAME_MAX_LENGTH
-from excel_handler import ExcelTemplateHandler
-        
+from treelib import Tree
+import typer
+import openpyxl
+from config import (BASE_TEMPLATE,SOURCE_TABLES,STAKEHOLDER_REPORTS,REGION_HIERARCHY_TABLE,GCP_PROJECT, BIGQUERY_ANALYTICS_DATASET, SHEET_DEFINITIONS, REPORTS)
+from google.cloud import bigquery
 
-from utils import sanitize_date_fields, sanitize_numeric_types#, create_required_folders, resolve_report_output_path
+from utils.file_utils import slugify
+from utils.data_utils import sanitize_date_fields, sanitize_numeric_types
+
+import os
+from copy import copy
+from datetime import datetime, timedelta,date
+from dateutil.relativedelta import relativedelta
+import pandas as pd
+import logging
+from utils.duckdb_utils import aggregate_kpi_data, query_yearly_kpi, query_monthly_kpi
+from utils.data_utils import build_region_hierarchy
+
+
+logger = logging.getLogger(__name__)
+logging.basicConfig(level=logging.INFO)
+
 
 class StakeholderType(Enum):
     MINISTERE = "ministere"
     DRAC = "drac"
 
-class SheetType(Enum):
-    KPIS = "kpis"
-    TOP = "top"
-    LEXIQUE = "lexique"
-
 @dataclass
 class Stakeholder:
-    """Stakeholder - either Ministère or DRAC."""
+    """Stakeholder - either Ministère (no children) or DRAC (region → academies & departments)."""
     type: StakeholderType
     name: str
     desired_reports: List[str] = field(default_factory=list)
-    
+    academy_tree: Optional[Tree] = None
+    department_tree: Optional[Tree] = None
+
     def __post_init__(self):
+        # Assign default reports
         if not self.desired_reports:
             self.desired_reports = list(STAKEHOLDER_REPORTS.get(self.type.value, []))
 
+        # Only DRAC stakeholders get trees
+        if self.type == StakeholderType.DRAC:
+            # Fetch hierarchy from BigQuery
+            hierarchy = build_region_hierarchy()
+
+            # Ensure region exists
+            if self.name not in hierarchy:
+                raise ValueError(f"❌ Region '{self.name}' not found in hierarchy")
+
+            scope = hierarchy[self.name]
+
+            # Build departement tree
+            self.department_tree = Tree()
+            self.department_tree.create_node("Departements", "root_dept")
+            for dep in scope["departements"]:
+                self.department_tree.create_node(dep, f"{self.name}_dep_{dep}", parent="root_dept")
+
+            # Build academy tree
+            self.academy_tree = Tree()
+            self.academy_tree.create_node("Academies", "root_acad")
+            for acad in scope["academies"]:
+                self.academy_tree.create_node(acad, f"{self.name}_acad_{acad}", parent="root_acad")
+
+        else:
+            self.academy_tree = None
+            self.department_tree = None
+        # if self.type == StakeholderType.DRAC:
+        #     typer.secho(f"➡️ DRAC '{self.name}' with adademy tree:\n{self.academy_tree.show(stdout=False)}", fg="cyan")
+        #     typer.secho(f"➡️ DRAC '{self.name}' with department tree:\n{self.department_tree.show(stdout=False)}", fg="cyan")
+
 @dataclass
 class Sheet:
-    """Represents a sheet within a report."""
-    type: SheetType  # kpis or top
-    name: Optional[str] = None
-    scale: Optional[str] = None  # national, regional, departmental, academie
-    scope: Optional[str] = None  # individual, collective
-    ds: Optional[str] = None  # consolidation date
-    source_table: Optional[str] = None  # individual_data, collective_data, top_individual_offer_data
-    format_config: Optional[Dict[str, Any]] = field(default_factory=dict)
-    
+    """Represents a sheet instance within a report."""
+    definition: str                    # key from SHEET_DEFINITIONS
+    template_tab: str                  # template tab name
+    worksheet: Any                     # openpyxl worksheet instance
+    filters: Dict[str, Any] = field(default_factory=dict)
+    tab_name: Optional[str] = None
+    stakeholder: Optional[Any] = None  # reference to Stakeholder
+    context: Optional[Dict[str, Any]] = None  # region, academy, department, etc.
+    layout: Optional[Dict[str, Any]] = None
 
-    def generate_sheet_name(self) -> str:
-        """Generate a descriptive sheet name."""
-        if self.type == SheetType.LEXIQUE:
-            self.name = "Lexique"
-            return "Lexique"
-        tab = ALL_TABLES.get(self.scope,{}).get("tab", "")
-        scope = self.scope or ""
-        scale = self.scale or ""
-        assert (tab != ""), f"Tab name missing for source_table: {self.source_table}"
-        assert (scope != ""), f"Scope missing for sheet with source_table: {self.source_table}"
-        assert (scale != ""), f"Scale missing for sheet with source_table: {self.source_table}"
-        if self.format_config.get("is_reference"):
-            mascu_scope = "Individuel" if self.scope == "individual" else ("Collectif" if self.scope == "collective" else "")
-            tab_name = f"Réf. {self.scale.capitalize()} - {mascu_scope}"
-
-        else:
-            tab_name = f"{tab} - {scale.capitalize()}"
-        self.name = tab_name[:TAB_NAME_MAX_LENGTH]  # Excel sheet name limit
-        print(f"Generated sheet name: {self.name}")
-        return tab_name
+    def get_dimension_context(self) -> Optional[Dict[str, str]]:
+        """
+        Resolve dimension context for DuckDB queries based on filters and context.
         
-    def get_table_name(self) -> str:
-        """Get the DuckDB table name for this sheet."""
-        return f"{self.source_table}"
-    
-    # def build_query(self, filters: Dict[str, str]) -> str:
-    #     """Build SQL query for this sheet's data."""
-    #     where_clauses = []
-    #     for key, value in filters.items():
-    #         where_clauses.append(f"{key} = '{value}'")
-        
-    #     where_clause = " AND ".join(where_clauses) if where_clauses else "1=1"
-        
-    #     return f"""
-    #     SELECT * FROM {self.get_table_name()}
-    #     WHERE {where_clause}
-    #     ORDER BY partition_month, kpi_name
-    #     """
-        
-    def build_query(self, filters: Dict[str, str]) -> str:
-        """Build SQL query for this sheet's data with proper filtering."""
-        where_clauses = ["1=1"]
-        
-        # Add base dimension filters
-        for key, value in filters.items():
-            if key and value:
-                where_clauses.append(f"{key} = '{value}'")
-        
-        # Add academy-specific filters if needed
-        if hasattr(self, 'academy_filter') and self.academy_filter:
-            where_clauses.append(f"academy_name = '{self.academy_filter}'")
-        
-        # Add department-specific filters if needed
-        if hasattr(self, 'department_filter') and self.department_filter:
-            where_clauses.append(f"department_name = '{self.department_filter}'")
-        
-        where_clause = " AND ".join(where_clauses)
-        
-        # Choose query template based on source table
-        if self.source_table and ("top_" in self.source_table or self.type == SheetType.TOP):
-            query_template = """
-            SELECT *
-            FROM {table}
-            WHERE {where_clause}
-            ORDER BY partition_month, dimension_name, dimension_value
-            """
-        else:
-            query_template = """
-            SELECT partition_month, updated_at, dimension_name, dimension_value,
-                kpi_name, numerator, denominator, kpi
-            FROM {table}
-            WHERE {where_clause}
-            ORDER BY partition_month, dimension_name, dimension_value, kpi_name
-            """
-        
-        return query_template.format(
-            table=f"{self.get_table_name()}_data",
-            where_clause=where_clause
-        )
-        
-    def _fill_with_data(self, conn: duckdb.DuckDBPyConnection, stakeholder: Stakeholder, ds: str):
-        """Fill this sheet with data for the stakeholder."""
-        # Move your existing sheet filling logic here
-        pass    
-    
-    def _apply_formatting(self):
-        """Apply formatting to the worksheet."""
-        pass
-        
-@dataclass
-class Report:
-    """Represents a complete report file."""
-    target_name: str  # Name of target stakeholder
-    target_type: StakeholderType
-    ds: str  # consolidation date
-    scopes: List[str] = field(default_factory=list)  # ["individual", "collective"]
-    scales: List[str] = field(default_factory=list)  # ["national", "regional", "departmental","academie"]
-    sheets: List[Sheet] = field(default_factory=list)
-    output_path: Optional[Path] = None
-
-    def __post_init__(self):
-        if not self.sheets:
-            self.sheets = []
-
-    def generate_sheets_from_config(self, report_type: str):
-        """Generate sheets based on REPORTS configuration."""
-        if report_type not in REPORTS:
-            raise ValueError(f"Unknown report type: {report_type}. Available: {list(REPORTS.keys())}")
-        
-        config = REPORTS[report_type]
-        self.scopes = config["scopes"]
-        self.scales = config["scales"]
-        self.sheets = []
-
-        if config.get("lexique", True):
-            self.sheets.append(Sheet(type=SheetType.LEXIQUE))
-            
-        # Generate main sheets for each scale/scope combination
-        for scope in self.scopes:
-            # Add reference scale sheets if specified
-            if "ref_scale" in config:
-                self._add_sheets_for_scale_scope(config["ref_scale"], scope, is_reference=True)
-            for scale in self.scales:
-                self._add_sheets_for_scale_scope(scale, scope)
-    
-    
-    def _add_sheets_for_scale_scope(self, scale: str, scope: str, is_reference: bool = False):
-        """Add sheets for a specific scale/scope combination."""
-        
-        # Add KPI sheet
-        source_table = scope #ALL_TABLES.get(scope,{}).get("table_name")
-        
-        self.sheets.append(Sheet(
-            type=SheetType.KPIS,
-            scale=scale,
-            scope=scope,
-            ds=self.ds,
-            source_table=source_table,
-            format_config={"is_reference": is_reference}
-        ))
-        
-        # Add TOP sheets from config: Not applicable for reference scales
-        if not is_reference and scope in TOP_TABLES_BY_SCOPE:
-            for table in TOP_TABLES_BY_SCOPE[scope]:
-                self.sheets.append(Sheet(
-                    type=SheetType.TOP,
-                    scale=scale,
-                    scope=scope,
-                    ds=self.ds,
-                    source_table=table,
-                    format_config={"is_reference": is_reference}
-                ))
-                
-    def generate_tab_names(self):
-        """Generate tab names for all sheets."""
-        for sheet in self.sheets:
-            sheet.generate_sheet_name()
-
-    # def get_output_filename(self) -> str:
-    #     """Generate output filename based on target and scales."""
-    #     if self.target_type == StakeholderType.MINISTERE:
-    #         return "rapport_national.xlsx"
-    #     elif "regional" in self.scales:
-    #         return f"résumé_{self.target_name}.xlsx"
-    #     elif len(self.scales) == 1:
-    #         return f"{self.scales[0]}_{self.target_name}.xlsx"
-    #     else:
-    #         return f"detail_{self.target_name}.xlsx"
-    
-    def get_data_filters(self) -> Dict[str, str]:
-        """Get data filters based on target type and scales."""
-        if self.target_type == StakeholderType.MINISTERE:
-            return {"dimension_name": "NAT"}
-        else:
-            # For DRAC, filter by region
-            return {"dimension_name": "REG", "dimension_value": self.target_name}
-        
-
-    def get_output_filename(self) -> str:
-        """Generate output filename based on target, context, and scales."""
-        if self.target_type == StakeholderType.MINISTERE:
-            return "rapport_national.xlsx"
-        
-        # For DRAC stakeholders
-        if hasattr(self, 'academy_filter') and self.academy_filter:
-            # Academy-specific report
-            from utils import slugify
-            return f"académie_{slugify(self.academy_filter)}.xlsx"
-        
-        elif hasattr(self, 'department_filter') and self.department_filter:
-            # Department-specific report
-            from utils import slugify
-            return f"départemental_{slugify(self.department_filter)}.xlsx"
-        
-        else:
-            # Regional summary report
-            return f"{self.target_name.replace(' ', '_')}.xlsx"
-
-    def generate_excel_report(self, conn: duckdb.DuckDBPyConnection, template_path: Path):
-        """Generate the actual Excel file using the template handler."""
-        
-        handler = ExcelTemplateHandler(template_path)
-        handler.load_template()
-        
-        # Get data filters for this stakeholder/context
-        filters = self.get_data_filters()
-        
-        # Apply context filters to sheets if needed
-        for sheet in self.sheets:
-            if hasattr(self, 'academy_filter'):
-                sheet.academy_filter = self.academy_filter
-            if hasattr(self, 'department_filter'):
-                sheet.department_filter = self.department_filter
-        
+        Returns:
+            Dict with 'name' (dimension name) and 'value' (dimension value) for DuckDB queries
+        """
         try:
-            # Fill each sheet with data
-            for sheet in self.sheets:
-                handler.fill_sheet_with_data(sheet, conn, filters)
+            scale = self.filters.get("scale", "national")
             
-            # Save the completed report
-            handler.save_report(self.output_path)
+            # Map scale to dimension name
+            dimension_name_map = {
+                "national": "NAT",
+                "region": "REG", 
+                "academie": "ACAD",
+                "departement": "DEP"
+            }
             
-        finally:
-            handler.close()
+            dimension_name = dimension_name_map.get(scale, "NAT")
+            
+            # Get dimension value from context
+            if scale == "national":
+                dimension_value = "NAT"
+            else:
+                dimension_value = self.context.get(scale) if self.context else None
+                
+            if not dimension_value and scale != "national":
+                logger.warning(f"Could not resolve dimension value for scale '{scale}' in context {self.context}")
+                return None
+                
+            return {
+                "name": dimension_name,
+                "value": dimension_value
+            }
+            
+        except Exception as e:
+            logger.warning(f"Failed to resolve dimension context: {e}")
+            return None
         
-# @dataclass
-# class TargetStakeholder:
-#     """Represents who gets a report and their filtering requirements."""
-#     name: str   # Ministrere, DRAC
-#     dimension_name: str  # NAT, REG
-#     dimension_value: str
-#     level: str  # national, regional, academie, departemental
-#     scope: str  # individual, collective
-#     region_name: Optional[str] = None
-#     academie_name: Optional[str] = None
-    
-#     def get_data_filters(self) -> Dict[str, str]:
-#         """Generate SQL WHERE clause filters."""
-#         filters = {"dimension_name": self.dimension_name}
-#         if self.dimension_value != self.dimension_name:  # Not national
-#             filters["dimension_value"] = self.dimension_value
-#         return filters
-    
-#     def get_output_filename(self) -> str:
-#         """Generate appropriate filename."""
-#         if self.level == "national":
-#             return "rapport_national.xlsx"
-#         elif self.level == "regional":
-#             return f"rapport_{self.dimension_value}.xlsx"
-#         elif self.level == "academie":
-#             return f"collectif_academie_{self.dimension_value}.xlsx"
-#         elif self.level == "departemental":
-#             return f"individuel_{self.dimension_value}.xlsx"
-#         return f"report_{self.name}.xlsx"
+    def __post_init__(self):
+        if not self.context:
+            self.context = self.get_dimension_context()
 
-# @dataclass
-# class Sheet:
-#     """Represents a sheet within a report."""
-#     type: str  # individuel, collectif, tops_individuel, tops_collectif
-#     worksheet: openpyxl.worksheet.worksheet.Worksheet
-#     kpi_filters: List[tuple]  # (row_idx, kpi_name, select_type, agg_type)
-    
-#     def fill_with_data(self, conn: duckdb.DuckDBPyConnection, stakeholder: TargetStakeholder, ds: str):
-#         """Fill this sheet with data for the stakeholder."""
-#         # Move your existing sheet filling logic here
-#         pass
-    
-#     def _apply_formatting(self):
-#         """Apply formatting to the worksheet."""
-#         pass
-
-# @dataclass
-# class Report:
-#     """Represents a complete report file."""
-#     stakeholder: TargetStakeholder
-#     base_template_path: Path
-#     output_path: Path
-#     scope: str = "all"   # individual, collective, all
-#     sheets: List[Sheet] = None
-#     workbook: Optional[openpyxl.Workbook] = None
-    
-#     def load_workbook_template(self):
-#         """Load the Excel workbook."""
-#         self.workbook = openpyxl.load_workbook(self.base_template_path)
-    
-#     def generate_sheets(self):
-#         """Initialize sheets based on the template."""
-#         if not self.workbook:
-#             self.load_workbook()
+    # def preprocess(self, ds: str):
+    #     """Placeholder for preprocessing - now handled by ReportOrchestrationService."""
+    #     pass
         
+    # def fill_data(self, duckdb_conn: duckdb.DuckDBPyConnection, ds: str):
+    #     """Placeholder for data filling - now handled by ReportOrchestrationService.""" 
+    #     pass
         
-#         self.sheets = []
-        
-#         for sheet_name in self.workbook.sheetnames:
-#             worksheet = self.workbook[sheet_name]
-#             # Determine type and kpi_filters based on sheet_name or content
-#             type = "individual"
-    
-#     def generate(self, conn: duckdb.DuckDBPyConnection, ds: str):
-#         """Generate the complete report."""
-#         if not self.workbook:
-#             self.load_workbook_template()
-        
-#         # Fill each sheet
-#         for sheet in self.sheets:
-#             sheet.fill_with_data(conn, self.stakeholder, ds)
-        
-#         # Apply final formatting and save
-#         self._finalize_and_save()
-    
-#     def _finalize_and_save(self):
-#         """Apply final formatting and save the report."""
-#         # Move lexique sheet first, hide columns, etc.
-#         self.output_path.parent.mkdir(parents=True, exist_ok=True)
-#         self.workbook.save(self.output_path)
+    # def postprocess(self):
+    #     """Placeholder for postprocessing."""
+    #     pass
 
 
 
+class Report:
+    """Represents a single Excel report for a stakeholder."""
+    
+    def __init__(self, report_type: str, stakeholder, base_template_path: Path, output_path: Path, context: Dict[str, Any] = None):
+        self.report_type = report_type
+        self.stakeholder = stakeholder
+        self.base_template_path = base_template_path
+        self.output_path = output_path
+        self.context = context or {}
+        self.workbook: Any = None          # openpyxl.Workbook
+        self.sheets: List[Sheet] = []      # list of Sheet instances
+
+    def _load_template(self):
+        """Load the template workbook directly as the report workbook."""
+        self.workbook = openpyxl.load_workbook(self.base_template_path)
+        
+        # Remove default sheet if it exists and is empty
+        if "Sheet" in self.workbook.sheetnames and len(self.workbook["Sheet"].rows) == 0:
+            self.workbook.remove(self.workbook["Sheet"])
+
+    def _build_tab_name(self, definition_key: str, filters: Dict[str, Any], context: Dict[str, Any]):
+        """Return a user-friendly tab name based on sheet definition, filters, and context."""
+        
+        if definition_key == "individual_kpis":
+            base = "Part Individuelle"
+            scale = filters.get("scale", "").capitalize()
+            return f"{base} ({scale})" if scale and scale != "National" else base
+        
+        elif definition_key == "collective_kpis":
+            base = "Part Collective" 
+            scale = filters.get("scale", "").capitalize()
+            return f"{base} ({scale})" if scale and scale != "National" else base
+    
+        else:
+            tab_name = SHEET_DEFINITIONS.get(definition_key, {}).get("tab_name")
+            if tab_name is None:
+                tab_name = SHEET_DEFINITIONS.get(definition_key, {}).get("template_tab", definition_key)
+            return tab_name.capitalize()
+    
+    def _resolve_sheets(self):
+        """Create Sheet objects from template tabs based on report blueprint."""
+        blueprint = REPORTS[self.report_type]
+        
+        for sheet_info in blueprint["sheets"]:
+            definition_key = sheet_info["definition"]
+            template_info = SHEET_DEFINITIONS[definition_key]
+            
+            # Decide tab name based on definition + filters + context
+            tab_name = self._build_tab_name(
+                definition_key, 
+                {**sheet_info.get("filters", {}), **self.context},
+                self.context
+            )
+
+            # Check if the tab already exists
+            if tab_name in self.workbook.sheetnames:
+                ws = self.workbook[tab_name]
+            else:
+                # Copy the template tab
+                template_ws = self.workbook[template_info["template_tab"]]
+                ws = self.workbook.copy_worksheet(template_ws)
+       
+            if len(tab_name) > 31:
+                typer.secho(f"⚠️ Tab name '{tab_name}' exceeds 31 characters, truncating.", fg="yellow")
+                tab_name = tab_name[:31] # truncate if too long
+            ws.title = tab_name  
+            
+            # Create Sheet instance
+            sheet = Sheet(
+                definition=definition_key,
+                template_tab=template_info["template_tab"],
+                worksheet=ws,
+                filters={**sheet_info.get("filters", {}), **self.context},
+                tab_name=tab_name,
+                stakeholder=self.stakeholder,
+                context=self.context
+            )
+            self.sheets.append(sheet)
+
+    def _cleanup_template_sheets(self, used_tabs: List[str]):
+        """Remove all template/unused sheets and enforce blueprint order."""
+        # Remove any sheet not in the used list
+        for sheet_name in list(self.workbook.sheetnames):
+            if sheet_name not in used_tabs:
+                self.workbook.remove(self.workbook[sheet_name])
+
+        # Enforce order from used_tabs
+        for idx, tab_name in enumerate(used_tabs):
+            ws = self.workbook[tab_name]
+            self.workbook._sheets.remove(ws)
+            self.workbook._sheets.insert(idx, ws)
+    
+    def build(self, ds: str, duckdb_conn: duckdb.DuckDBPyConnection):
+        """
+        Build the report workbook using the new service architecture.
+        
+        Args:
+            ds: Consolidation date in YYYY-MM-DD format
+            duckdb_conn: DuckDB connection for data queries
+        """
+        from services.orchestration import ReportOrchestrationService
+
+        typer.secho(f"➡️ Building report: {self.report_type} for {self.stakeholder.name}", fg="cyan")
+        
+        # Step 1: Load template and create sheets
+        self._load_template()
+        self._resolve_sheets()
+        self._cleanup_template_sheets([sheet.tab_name for sheet in self.sheets])
+        
+        # Step 2: Use ReportOrchestrationService for complex processing
+        orchestrator = ReportOrchestrationService(duckdb_conn)
+        stats = orchestrator.process_all_sheets(self.sheets, ds)
+        
+        # Step 3: Log final statistics
+        if stats["sheets_successful"] == len(self.sheets):
+            typer.secho(f"✅ Report completed successfully: {self.output_path.name}", fg="green")
+        elif stats["sheets_successful"] > 0:
+            typer.secho(
+                f"⚠️ Report completed with warnings: {stats['sheets_successful']}/{len(self.sheets)} sheets successful", 
+                fg="yellow"
+            )
+        else:
+            typer.secho("❌ Report failed: No sheets processed successfully", fg="red")
+
+    def save(self):
+        """Save the workbook to the output path."""
+        if not self.output_path.suffix == ".xlsx":
+            self.output_path = self.output_path.with_suffix(".xlsx")
+        self.output_path.parent.mkdir(parents=True, exist_ok=True)
+        self.workbook.save(self.output_path)
+        typer.echo(f"💾 Saved report: {self.output_path}")
+        
+class ReportPlanner:
+    """Expands a stakeholder into the list of reports they should get."""
+
+    def __init__(self, stakeholder):
+        self.stakeholder = stakeholder
+
+    def plan_reports(self) -> List[Dict[str, Any]]:
+        """
+        Returns a list of report jobs:
+        {
+            "report_type": str,
+            "context": dict,  # region, academy, department, etc.
+            "output_path": Path
+        }
+        """
+        jobs = []
+
+        if self.stakeholder.type.name == "MINISTERE":
+            # Ministère gets only the national summary
+            # base_path = Path("NATIONAL")
+            jobs.append(
+                {
+                    "report_type": "national_summary",
+                    "context": {},
+                    "output_path": "rapport_national.xlsx"
+                }
+            )
+            return jobs
+
+        # DRAC gets:
+        # 1) Regional summary (includes national for comparison)
+        region_name = self.stakeholder.name
+        # base_path = Path("REGIONAL") / f"{region_name}"
+        jobs.append(
+            {
+                "report_type": "region_summary",
+                "context": {"region": region_name},
+                "output_path": "rapport_regional.xlsx"
+            }
+            )
+
+        # 2) One detailed report per académie (collective scope)
+        if self.stakeholder.academy_tree:
+            for node in self.stakeholder.academy_tree.all_nodes():
+                if node.tag not in ["root_acad", "Academies"]:
+                    jobs.append(
+                        {
+                            "report_type": "academy_detail",
+                            "context": {"region": region_name, "academie": node.tag},
+                            "output_path": f"academie_{node.tag}.xlsx"
+                        }
+                    )
+
+        # 3) One detailed report per département (individual scope)
+        if self.stakeholder.department_tree:
+            for node in self.stakeholder.department_tree.all_nodes():
+                if node.tag not in ["root_dept", "Departements"]:
+                    jobs.append(
+                        {
+                            "report_type": "department_detail",
+                            "context": {"region": region_name, "departement": node.tag},
+                            "output_path": f"departement_{node.tag}.xlsx"
+                        }
+                    )
+
+        return jobs
+    
 
 class ExportSession:
     """Manages the entire export session with proper resource cleanup."""
-    
-    def __init__(self, ds: str, template_path: Path, output_dir: Path, scope: str, target: str):
+    def __init__(self, ds: str):
         self.ds = ds
-        self.template_path = template_path
-        self.output_dir = output_dir
-        self.scope: str = scope
-        self.target = target
-        self.conn: Optional[duckdb.DuckDBPyConnection] = None
-        self.hierarchy_data = None
-        
-        
+        self.tables_to_load = SOURCE_TABLES
+        self.conn: Optional[duckdb.duckdb.DuckDBPyConnection] = None
+
     def __enter__(self):
-        self.output_dir.mkdir(parents=True, exist_ok=True)
         return self
         
     def __exit__(self, exc_type, exc_val, exc_tb):
         if self.conn:
             self.conn.close()
-    
-    def validate_inputs(self, target: str, scope: str):
-        """Validate all inputs before starting."""
-        if target.lower() not in ("national", "drac", "all"):
-            raise ValueError(f"Invalid target: {target}")
-        if scope not in ("individual", "collective", "all"):
-            raise ValueError(f"Invalid scope: {scope}")
-        if not self.template_path.exists():
-            raise FileNotFoundError(f"Template not found: {self.template_path}")
+
     
     def load_data(self):
         """Load tables based on self.scope."""
-        if self.scope not in SCOPE_TABLES:
-            raise ValueError(f"Unknown scope: {self.scope}. Available: {list(SCOPE_TABLES.keys())}")
-        
-        tables_to_load = SCOPE_TABLES[self.scope]
         
         client = bigquery.Client(project=GCP_PROJECT)
         self.conn = duckdb.connect(":memory:")
         
-        for table_key in tables_to_load:
-            if table_key not in QUERIES:
-                typer.echo(f"Warning: No query defined for '{table_key}', skipping")
-                continue
-                
-            self._load_table(client, table_key)
+        for _, config in self.tables_to_load.items():
+            self._load_table(client,config)
     
-    def _load_table(self, client: bigquery.Client, table_key: str):
+    def _load_table(self, client: bigquery.Client, config: Dict):
         """Load a single table into DuckDB."""
-        query = QUERIES[table_key]
-        table_name = f"{table_key}_data"
-        
-        typer.echo(f"Loading {table_key} from BigQuery...")
-        
-        
+        table_name = config['table']
+        dataset = config.get('dataset', BIGQUERY_ANALYTICS_DATASET)
+        project_id = config.get('project', GCP_PROJECT)
+        typer.echo(f"Loading {table_name} from BigQuery...")
+        query = f"SELECT * FROM `{project_id}.{dataset}.{table_name}`"
         with warnings.catch_warnings():
             warnings.filterwarnings("ignore", message=".*BigQuery Storage module not found.*")
             df = client.query(query).to_dataframe()
-            typer.echo(f"{table_key} dtypes: {df.dtypes}")
+            typer.echo(f"{table_name} dtypes: {df.dtypes}")
         if df.empty:
-            typer.echo(f"Warning: {table_key} returned no data")
+            typer.echo(f"Warning: {table_name} returned no data")
             return
         
         df = sanitize_date_fields(df, "partition_month")
         df = sanitize_numeric_types(df)
             
-        typer.echo(f"Loaded {len(df):,} records for {table_key}")
+        typer.echo(f"Loaded {len(df):,} records for {table_name}")
         
         # Register and create table
-        self.conn.register(f"{table_key}_tmp", df)
-        self.conn.execute(f"CREATE TABLE {table_name} AS SELECT * FROM {table_key}_tmp")
+        self.conn.register(f"{table_name}_tmp", df)
+        self.conn.execute(f"CREATE TABLE {table_name} AS SELECT * FROM {table_name}_tmp")
         
         # Create indexes based on available columns
         self._create_indexes(table_name, df.columns.tolist())
-        typer.echo(f"MOCK -> Created {table_name} table with indexes")
+        typer.echo(f"Created {table_name} table with indexes")
 
     
     def _create_indexes(self, table_name: str, columns: List[str]):
-        """Create indexes on commonly queried columns."""
-        pass
-    
-    def create_stakeholders(self, target: str) -> List[Stakeholder]:
-        """Create stakeholder list based on target - enhanced for multiple DRAC reports."""
-        stakeholders = []
+        """Create indexes on commonly queried columns with error recovery."""
+        indexes_created = []
+        indexes_failed = []
         
-        if target in ("national", "all"):
-            stakeholders.append(Stakeholder(
-                type=StakeholderType.MINISTERE,
-                name="Ministere"
-            ))
+        typer.secho(f"🔍 Creating indexes for table: {table_name}", fg="cyan")
         
-        if target in ("drac", "all"):
-            # Load hierarchy if not already loaded
-            if not self.hierarchy_data:
-                from utils import build_region_hierarchy
-                self.hierarchy_data = build_region_hierarchy()
-                
-            for region_name, region_data in self.hierarchy_data.items():
-                # 1. Create base DRAC stakeholder for regional summary
-                regional_stakeholder = Stakeholder(
-                    type=StakeholderType.DRAC,
-                    name=region_name,
-                    desired_reports=["regional_summary"]  # Global regional report
+        # List of indexes to attempt
+        index_operations = []
+        
+        # Core filtering columns
+        core_filter_columns = ['dimension_name', 'dimension_value', 'partition_month']
+        kpi_columns = ['kpi_name']
+        
+        if all(col in columns for col in core_filter_columns):
+            index_operations.append({
+                "name": "core_filters",
+                "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_core_filters ON {table_name} (dimension_name, dimension_value, partition_month)",
+                "description": "Core filters composite index"
+            })
+        
+        if all(col in columns for col in core_filter_columns + kpi_columns):
+            index_operations.append({
+                "name": "kpi_filters", 
+                "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_kpi_filters ON {table_name} (kpi_name, dimension_name, dimension_value, partition_month)",
+                "description": "KPI-specific composite index"
+            })
+        
+        # Individual indexes
+        for col in ['partition_month', 'dimension_name','dimension_value', 'kpi_name']:
+            if col in columns:
+                index_operations.append({
+                    "name": col,
+                    "sql": f"CREATE INDEX IF NOT EXISTS idx_{table_name}_{col} ON {table_name} ({col})",
+                    "description": f"Individual index on {col}"
+                })
+        
+
+        # Execute each index creation with individual error handling
+        for idx_op in index_operations:
+            try:
+                self.conn.execute(idx_op["sql"])
+                indexes_created.append(idx_op["name"])
+                typer.echo(f"   ✓ {idx_op['description']}")
+            except Exception as e:
+                indexes_failed.append(idx_op["name"])
+                if "INTERNAL Error" in str(e):
+                    typer.secho(f"   ⚠ {idx_op['description']} - skipped (DuckDB internal error)", fg="yellow")
+                else:
+                    typer.secho(f"   ❌ {idx_op['description']} - failed: {str(e)[:50]}...", fg="red")
+        
+        # Summary
+        if indexes_created:
+            typer.secho(f"📊 {table_name}: Created {len(indexes_created)} indexes successfully", fg="green")
+        if indexes_failed:
+            typer.secho(f"⚠ {table_name}: {len(indexes_failed)} indexes failed (table will still work)", fg="yellow")
+        
+
+    def process_stakeholder(self, stakeholder_type: str, name: str, output_path: Path, ds: str):
+        """Process reports for a given stakeholder using new service architecture."""
+        stakeholder = Stakeholder(
+            type=StakeholderType.MINISTERE if stakeholder_type == "ministere" else StakeholderType.DRAC,
+            name="Ministère" if stakeholder_type == "ministere" else name,
+        )
+        typer.secho(f"➡️ Processing reports for {stakeholder.type.value} - {stakeholder.name}", fg="cyan")
+        
+        # Generate reports based on stakeholder's desired reports
+        planner = ReportPlanner(stakeholder)
+        report_jobs = planner.plan_reports()
+        
+        successful_reports = 0
+        total_reports = len(report_jobs)
+        
+        typer.secho(f"📋 Planning to generate {total_reports} reports", fg="cyan")
+        
+        for job in report_jobs:
+            try:
+                # Create Report instance
+                report = Report(
+                    report_type=job["report_type"],
+                    stakeholder=stakeholder,
+                    base_template_path=BASE_TEMPLATE,
+                    output_path=output_path / job["output_path"],
+                    context=job["context"]
                 )
-                stakeholders.append(regional_stakeholder)
                 
-                # 2. Create academy-specific stakeholders for collective reports
-                for academy in region_data["academies"]:
-                    academy_stakeholder = Stakeholder(
-                        type=StakeholderType.DRAC,
-                        name=region_name,
-                        desired_reports=["academie_detail"]
-                    )
-                    # Add academy context for filtering and naming
-                    academy_stakeholder.academy_name = academy
-                    academy_stakeholder.context_type = "academy"
-                    stakeholders.append(academy_stakeholder)
+                # Build and save report using new architecture
+                report.build(ds, self.conn)
+                report.save()
                 
-                # 3. Create department-specific stakeholders for individual reports  
-                for dept in region_data["departements"]:
-                    dept_stakeholder = Stakeholder(
-                        type=StakeholderType.DRAC, 
-                        name=region_name,
-                        desired_reports=["departemental_detail"]
-                    )
-                    # Add department context for filtering and naming
-                    dept_stakeholder.department_name = dept
-                    dept_stakeholder.context_type = "department"
-                    stakeholders.append(dept_stakeholder)
-        
-        return stakeholders
-    # def create_stakeholders(self, target: str) -> List[Stakeholder]:
-    #     """Create stakeholder list based on target."""
-    #     stakeholders = []
-        
-    #     if target in ("national", "all"):
-    #         stakeholders.append(Stakeholder(
-    #             type=StakeholderType.MINISTERE,
-    #             name="Ministere"
-    #         ))
-        
-    #     if target in ("drac", "all"):
-    #         # Load hierarchy if not already loaded
-    #         if not self.hierarchy_data:
-    #             from utils import build_region_hierarchy
-    #             self.hierarchy_data = build_region_hierarchy()
+                successful_reports += 1
                 
-    #         for region_name in self.hierarchy_data.keys():
-    #             stakeholders.append(Stakeholder(
-    #                 type=StakeholderType.DRAC,
-    #                 name=region_name
-    #             ))
+            except Exception as e:
+                typer.secho(f"❌ Failed to generate report {job['output_path']}: {e}", fg="red")
+                logger.error(f"Report generation failed for {job['output_path']}: {e}")
+                # Continue with next report instead of failing completely
         
-    #     return stakeholders
-
-
-
-if __name__ == "__main__":
-    report = Report(target_name="ile-de-france", target_type=StakeholderType.DRAC, ds="2025-08-01")
-    report.generate_sheets_from_config("academie_detail")
-    print(f"Generated {len(report.sheets)} sheets for report.")
-    print(f"Generated {report.sheets} sheets for report.")
-
-     
-    # def create_stakeholders(self, target: str) -> List[TargetStakeholder]:
-    #     """Create stakeholder list based on target."""
-    #     stakeholders = []
+        # Final summary
+        if successful_reports == total_reports:
+            typer.secho(f"✅ All {total_reports} reports generated successfully for {stakeholder.name}", fg="green")
+        elif successful_reports > 0:
+            typer.secho(
+                f"⚠️ {successful_reports}/{total_reports} reports generated successfully for {stakeholder.name}", 
+                fg="yellow"
+            )
+        else:
+            typer.secho(f"❌ No reports generated successfully for {stakeholder.name}", fg="red")
         
-    #     if target in ("national", "all"):
-    #         stakeholders.append(TargetStakeholder(
-    #             name="Ministere",
-    #             dimension_name="NAT",
-    #             dimension_value="NAT",
-    #             level="national",
-    #             scopes=["individual", "collective"]
-    #         ))
-        
-    #     if target in ("regional", "departemental", "all"):
-    #         # Load hierarchy if not already loaded
-    #         if not self.hierarchy_data:
-    #             from utils.hierarchy import build_region_hierarchy
-    #             self.hierarchy_data = build_region_hierarchy()
-            
-    #         # Create regional stakeholders
-    #         for region_name in self.hierarchy_data.keys():
-    #             if target in ("regional", "all"):
-    #                 stakeholders.append(TargetStakeholder(
-    #                     name=f"DRAC {region_name}",
-    #                     dimension_name="REG",
-    #                     dimension_value=region_name,
-    #                     level="regional",
-    #                     scopes=["individual", "collective"],
-    #                     region_name=region_name
-    #                 ))
-                
-    #             # Create departmental stakeholders
-    #             if target in ("departemental", "all"):
-    #                 hierarchy = self.hierarchy_data[region_name]
-    #                 for academy, departments in hierarchy.get("academy_departments", {}).items():
-    #                     for department in departments:
-    #                         stakeholders.append(TargetStakeholder(
-    #                             name=f"Department {department}",
-    #                             dimension_name="DEP",
-    #                             dimension_value=department,
-    #                             level="departemental",
-    #                             scope="individual",
-    #                             region_name=region_name,
-    #                             academie_name=academy
-    #                         ))
-        
-    #     return stakeholders
-    
-    # def create_reports(self, stakeholder: List[TargetStakeholder], scope: str) -> List[Report]:
-    #     """Create report objects for all stakeholders."""
-    #     reports = []
-        
-    #     if scope != "all" and scope != stakeholder.scope:
-    #         continue
-        
-    #     # Create output path
-    #     output_path = self._get_output_path(stakeholder,scope)
-        
-    #     # Create report
-    #     report = Report(
-    #         stakeholder=stakeholder,
-    #         template_path=self.template_path,
-    #         output_path=output_path,
-    #         scope = scope,
-    #         sheets=[]  # Will be populated when workbook is loaded
-    #     )
-        
-    #     reports.append(report)
-        
-    #     return reports
-    
-    # def _get_output_path(self, stakeholder: TargetStakeholder) -> Path:
-    #     """Generate output path for stakeholder."""
-    #     dated_dir = get_dated_base_dir(self.output_dir, self.ds)
-        
-    #     if stakeholder.level == "national":
-    #         return dated_dir / "NATIONAL" / stakeholder.get_output_filename()
-    #     elif stakeholder.level == "regional":
-    #         return dated_dir / "REGIONAL" / stakeholder.region_name / stakeholder.get_output_filename()
-    #     elif stakeholder.level == "departemental":
-    #         return dated_dir / "REGIONAL" / stakeholder.region_name / "ACADEMIES" / stakeholder.academie_name / stakeholder.get_output_filename()
-        
-    #     return dated_dir / stakeholder.get_output_filename()
-    
-    # def generate_all_reports(self, reports: List[Report]):
-    #     """Generate all reports with progress tracking."""
-    #     total = len(reports)
-        
-    #     for i, report in enumerate(reports, 1):
-    #         try:
-    #             print(f"[{i}/{total}] Generating {report.stakeholder.name}...")
-    #             report.generate(self.conn, self.ds)
-    #             print(f"    ✅ Saved to {report.output_path}")
-    #         except Exception as e:
-    #             print(f"    ❌ Failed: {e}")
-    #             continue
