+++ jobs/etl_jobs/external/ppg/utils/duckdb_utils.py
