--- jobs/etl_jobs/external/ppg2/pyproject.toml
@@ -1,38 +1,0 @@
-## ruff
-# Recommended ruff config for now, to be updated as we go along.
-[tool.ruff]
-target-version = 'py310'
-
-lint.select = ["E4", "E7", "E9", "F", "I"]
-# TARGET : lint.select = [
-#     "E",      # pycodestyle
-#     "W",      # pycodestyle
-#     "F",      # Pyflakes
-#     "B",      # flake8-bugbear
-#     "C4",     # flake8-comprehensions
-#     "I",      # isort
-#     "SIM",    # flake8-simplify
-#     "TCH",    # flake8-type-checking
-#     "TID",    # flake8-tidy-imports
-#     "UP",     # pyupgrade
-#     "PT",     # flake8-pytest-style
-#     "RUF",    # Ruff-specific rules
-#     "FBT001", # flake8-boolean-trap
-# ]
-
-[tool.ruff.lint.pydocstyle]
-# Automatically disable rules that are incompatible with Google docstring convention
-convention = "google"
-
-[tool.ruff.lint.pycodestyle]
-max-doc-length = 100
-
-[tool.ruff.lint.flake8-tidy-imports]
-ban-relative-imports = "all"
-
-[tool.ruff.lint.flake8-type-checking]
-strict = true
-runtime-evaluated-base-classes = ["pydantic.BaseModel"]
-# Pydantic needs to be able to evaluate types at runtime
-# see https://pypi.org/project/flake8-type-checking/ for flake8-type-checking documentation
-# see https://beta.ruff.rs/docs/settings/#flake8-type-checking-runtime-evaluated-base-classes for ruff documentation
